+++ CHANGES.rst
@@ -10,6 +10,9 @@ Changelog
 2.1.9 (2015-09-07)
 ------------------
 
+- Fix modeleditor for Plone 5
+  [vangheem]
+
 - Restrict allowed field types to regular types (so plone.app.users do not show
   up)
   [ebrehault]